Slightly improved flow.
Changed "I'm just gonna brute this!" To "I'm going in!" I feel like this makes it flow a little better.

Also, changed "Um, it's a late hour" to "Um, it's late"

Changed "Right. It's Eva's Curse." to "Right. It's the Eva's Curse.

Changed "But I saved her at that time!" to "But I saved her!"

Changed "The outcome was finding you" to "We only found you."