Turns out "Mk." appears onscreen after all!
Also restoring the full meaning to 00:47:01, which I had omitted completely by accident.

I do see that a question mark was added to 01:09:56 in the Japanese transcription, but the translation doesn't really seem to make any sense as a question... unless I'm missing something. So I've changed it back for now.